refactoring: Create module resources
